pyproject.toml
[tool.poetry]
name = "omogenjudge"
version = "0.1.0"
description = ""
authors = ["Johan Sannemo <jsannemo@jsannemo.se>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
psycopg2 = "^2.9.1"
problemtools = {git = "https://github.com/kattis/problemtools", branch = "develop"}
django-crispy-forms = "^1.12.0"
crispy-bootstrap5 = "^0.4"
django-debug-toolbar = "^3.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
